+++ doc/examples/preproc_with_topology.py
@@ -1,10 +1,11 @@
 from fluidimage import get_path_image_samples
-from fluidimage.topologies.preproc import TopologyPreproc
+from fluidimage.preproc import Topology
 
-params = TopologyPreproc.create_default_params()
+params = Topology.create_default_params()
 
 params.preproc.series.path = get_path_image_samples() / "Jet/Images"
-params.preproc.series.str_subset = "i+60:i+62, 0"
+params.preproc.series.str_subset = "i,:"
+params.preproc.series.ind_start = 60
 
 print("Available preprocessing tools: ", params.preproc.tools.available_tools)
 params.preproc.tools.sequence = [
@@ -25,7 +26,7 @@ params.preproc.tools.temporal_median.window_shape = (5, 2, 2)
 params.preproc.tools.global_threshold.enable = True
 params.preproc.tools.global_threshold.minima = 0.0
 
-topology = TopologyPreproc(params, logging_level="info")
+topology = Topology(params, logging_level="info", nb_max_workers=4)
 
 # Compute in parallel
 topology.compute()